+++ fixtures/custom-article-template-app/config.rb
@@ -1,0 +1,3 @@
+activate :blog do |blog|
+  blog.new_article_template = 'my_custom_article.tt'
+end